test: fix failing test
